move firmwares endpoints to its own controllers
