+++ .eslintrc
@@ -1,37 +1,14 @@
 {
-  "extends": [
-    "airbnb-base",
-    "prettier"
-  ],
+  "extends": ["airbnb-base", "prettier"],
   "parser": "babel-eslint",
-  "plugins": [
-    "flowtype",
-    "sorting"
-  ],
+  "plugins": ["flowtype", "sorting"],
   "rules": {
     "class-methods-use-this": 0,
     "eol-last": 2,
     "flowtype/define-flow-type": 1,
-    "flowtype/require-parameter-type": 1,
-    "flowtype/require-valid-file-annotation": [
-      2,
-      "always"
-    ],
-    "flowtype/require-return-type": [
-      1,
-      "always",
-      {
-        "annotateUndefined": "never"
-      }
-    ],
-    "flowtype/space-after-type-colon": [
-      1,
-      "always"
-    ],
-    "flowtype/space-before-type-colon": [
-      1,
-      "never"
-    ],
+    "flowtype/require-valid-file-annotation": [2, "always"],
+    "flowtype/space-after-type-colon": [1, "always"],
+    "flowtype/space-before-type-colon": [1, "never"],
     "flowtype/use-flow-type": 1,
     "import/first": 0,
     "import/newline-after-import": 0,
